Renaming mitmath -> calc
